Components are now separated into one component per file.
