+++ src/components/Board.tsx
@@ -1,0 +1,42 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "../index.css";
+import { Square } from "./Square"
+
+interface BoardProps {
+    squares: Array<string | null>;
+    onClick: (i: number) => void;
+}
+
+export class Board extends React.Component<BoardProps> {
+    renderSquare(i: number) {
+        return (
+            <Square
+                value={this.props.squares[i]}
+                onClick={() => this.props.onClick(i)}
+            />
+        );
+    }
+
+    render() {
+        return (
+            <div>
+                <div className="board-row">
+                    {this.renderSquare(0)}
+                    {this.renderSquare(1)}
+                    {this.renderSquare(2)}
+                </div>
+                <div className="board-row">
+                    {this.renderSquare(3)}
+                    {this.renderSquare(4)}
+                    {this.renderSquare(5)}
+                </div>
+                <div className="board-row">
+                    {this.renderSquare(6)}
+                    {this.renderSquare(7)}
+                    {this.renderSquare(8)}
+                </div>
+            </div>
+        );
+    }
+}
