+++ src/components/Square.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "../index.css";
+
+interface SquareProps {
+  value: string | null;
+  onClick: () => void;
+}
+
+export function Square(props: SquareProps) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+
