+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { Square } from "./Square"
+export { Board } from "./Board"
+export { Game } from "./Game"
